+++ settings/dev.py
@@ -8,6 +8,14 @@ from task_processor.task_run_method import TaskRunMethod
 
 env = Env()
 
+TEMPLATES = [
+    {
+        "BACKEND": "django.template.backends.django.DjangoTemplates",
+        "DIRS": ["templates"],
+        "APP_DIRS": True,
+    },
+]
+
 # Settings expected by `mypy_django_plugin`
 AWS_SES_REGION_ENDPOINT: str
 SEGMENT_RULES_CONDITIONS_LIMIT: int
@@ -20,10 +28,17 @@ DATABASES = {
     "default": dj_database_url.parse(
         env(
             "DATABASE_URL",
-            default="postgresql://postgres:password@localhost:5432/flagsmith",
-        )
-    )
+            default="postgresql://postgres@localhost:5432/postgres",
+        ),
+    ),
+    "task_processor": dj_database_url.parse(
+        env(
+            "TASK_PROCESSOR_DATABASE_URL",
+            default="postgresql://postgres@localhost:5433/postgres",
+        ),
+    ),
 }
+TASK_PROCESSOR_DATABASES = ["default"]
 INSTALLED_APPS = [
     "django.contrib.auth",
     "django.contrib.contenttypes",
@@ -48,7 +63,8 @@ TASK_DELETE_INCLUDE_FAILED_TASKS = False
 TASK_DELETE_RETENTION_DAYS = 15
 TASK_DELETE_RUN_EVERY = timedelta(days=1)
 TASK_DELETE_RUN_TIME = time(5, 0, 0)
-TASK_PROCESSOR_MODE = False
+TASK_PROCESSOR_MODE = env.bool("RUN_BY_PROCESSOR", default=False)
+DOCGEN_MODE = env.bool("DOCGEN_MODE", default=False)
 TASK_RUN_METHOD = TaskRunMethod.TASK_PROCESSOR
 
 # Avoid models.W042 warnings