+++ src/common/core/main.py
@@ -5,7 +5,11 @@ import sys
 import tempfile
 import typing
 
-from django.core.management import execute_from_command_line
+from django.core.management import (
+    execute_from_command_line as django_execute_from_command_line,
+)
+
+from common.core.cli import healthcheck
 
 logger = logging.getLogger(__name__)
 
@@ -48,6 +52,9 @@ def ensure_cli_env() -> typing.Generator[None, None, None]:
         )
         os.environ["PROMETHEUS_MULTIPROC_DIR"] = prometheus_multiproc_dir_name
 
+    if "docgen" in sys.argv:
+        os.environ["DOCGEN_MODE"] = "true"
+
     if "task-processor" in sys.argv:
         # A hacky way to signal we're not running the API
         os.environ["RUN_BY_PROCESSOR"] = "true"
@@ -56,7 +63,25 @@ def ensure_cli_env() -> typing.Generator[None, None, None]:
         yield
 
 
-def main() -> None:
+def execute_from_command_line(argv: list[str]) -> None:
+    try:
+        subcommand = argv[1]
+        subcommand_main = {
+            "healthcheck": healthcheck.main,
+            # Backwards compatibility for task-processor health checks
+            # See https://github.com/Flagsmith/flagsmith-task-processor/issues/24
+            "checktaskprocessorthreadhealth": healthcheck.main,
+        }[subcommand]
+    except (IndexError, KeyError):
+        django_execute_from_command_line(argv)
+    else:
+        subcommand_main(
+            argv[2:],
+            prog=f"{os.path.basename(argv[0])} {subcommand}",
+        )
+
+
+def main(argv: list[str] = sys.argv) -> None:
     """
     The main entry point to the Flagsmith application.
 
@@ -72,5 +97,5 @@ def main() -> None:
     `flagsmith <command> [options]`
     """
     with ensure_cli_env():
-        # Run Django
-        execute_from_command_line(sys.argv)
+        # Run own commands and Django
+        execute_from_command_line(argv)