+++ src/common/test_tools/__init__.py
@@ -1,3 +1,6 @@
-from common.test_tools.types import AssertMetricFixture
+from common.test_tools.types import AssertMetricFixture, SnapshotFixture
 
-__all__ = ("AssertMetricFixture",)
+__all__ = (
+    "AssertMetricFixture",
+    "SnapshotFixture",
+)